fix(world-protocol): native 54261 flat uint16 opcodes + uint16 wire (brick B re-value)
Re-value OpcodeClient/OpcodeServer from the authoritative HermesProxy
54261 enum (flat uint16 values; name=value), replacing the wrong
grouped 0xGGGGIIII uint32 scheme imported from TrinityCore 61581.

- Opcodes.h/.cpp: flat handler tables indexed directly by opcode value;
  drop GetOpcodeArrayIndex + NUM_CMSG/SMSG_OPCODES; add
  NUM_OPCODE_HANDLERS = 0x4000 (> max 54261 value 14943 + 1).
- WorldGlobals: AntiDos policy array sized/indexed by flat opcode value.
- WorldSocket: 2-byte (uint16) wire opcode read/write; IncomingPacketHeader
  EncryptedOpcode uint32->uint16; handle CMSG_LOG_DISCONNECT (pre-auth
  no-op); zero-init outbound GCM Tag; remove temp WORLDDBG diagnostic.

Handshake set verified byte-exact (CMSG_AUTH_SESSION=14181,
SMSG_AUTH_CHALLENGE=12360, etc.). No intra-enum value collisions.

+++ src/server/game/Server/WorldSocket.h
@@ -79,11 +79,12 @@ struct PacketHeader
     bool IsValidSize() const { return Size < 0x10000; }
 };
 
-// 3.4.3 modern wire header (inbound, 20 bytes). EncryptedOpcode is the first 4 bytes of the
-// encrypted payload, peeked forward so the body decrypt stays contiguous.
+// 3.4.3.54261 modern wire header (inbound, 18 bytes). The 54261 client uses FLAT uint16 opcodes,
+// so EncryptedOpcode is the first 2 bytes of the encrypted payload, peeked forward so the body
+// decrypt stays contiguous.
 struct IncomingPacketHeader : PacketHeader
 {
-    uint32 EncryptedOpcode;
+    uint16 EncryptedOpcode;
 };
 
 #pragma pack(pop)
